# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
        "/_monitoring/health": {
            "get": {
                "description": "Get App Health",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Monitoring"
                ],
                "summary": "Get App Health",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/shared.AppStatus"
                        }
                    }
                }
            }
        },
        "/_monitoring/info": {
            "get": {
                "description": "Get App Info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Monitoring"
                ],
                "summary": "Get App Info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/shared.AppInformation"
                        }
                    }
                }
            }
        },
        "/todos/{todo}": {
            "get": {
                "description": "Get A Todo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cargo Desi"
                ],
                "summary": "Get A Todo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "fileName",
                        "name": "todo",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/todo.GetTodoResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "shared.AppInformation": {
            "type": "object",
            "properties": {
                "appName": {
                    "type": "string"
                },
                "env": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "shared.AppStatus": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "todo.GetTodoResponse": {
            "type": "object",
            "properties": {
                "completed": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        }
    }
}